enrcrypt_with_password/requirements.txt
pycryptodome==3.20.0
tk==0.1.0